v1/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
scipy>=1.10.0
numpy>=1.24.0

# Model and training
peft>=0.4.0
accelerate>=0.20.0
bitsandbytes>=0.41.0  # 量子化サポート
# flash-attn>=2.3.0  # Flash Attention 2 - 別途インストールが必要（下記参照）

# Data loading and processing
datasets>=2.12.0
evaluate>=0.4.0  # Hugging Face評価ライブラリ

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0  # インタラクティブな可視化

# Configuration and utilities
pyyaml>=6.0
hydra-core>=1.3.0  # 設定管理
tqdm>=4.65.0
scikit-learn>=1.3.0

# Testing
pytest>=7.3.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.15.0

# Optional: Code evaluation
# human-eval>=1.0.0  # HumanEval評価用

# Flash Attention 2のインストール（オプション、高速化のため）:
# pip install flash-attn --no-build-isolation
# または、事前ビルド版を使用:
# pip install flash-attn --no-build-isolation --no-deps

# 環境変数読み込み
python-dotenv>=1.0.0
